Create README.md
This is a small task describing how to import the dataset and apply pre-processing to remove the duplicates, NULL values, and, unnecessary values. as we show in the result Beverage_Category and Calories that Signature Espresso Drinks have the highest calories. and the result between Beverage_Category and Sugars show that Signature Espresso Drinks have the highest sugars.